update offset to args
